references/agentic-framework-main/tools/kautilya/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kautilya"
version = "1.0.0"
description = "Interactive CLI for Enterprise Agentic Framework"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Architecture Team", email = "arch@company.com" }
]
keywords = ["cli", "agent", "framework", "orchestration", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # CLI Core
    "click>=8.1.7",
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.43",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "httpx>=0.26.0",
    "jinja2>=3.1.2",
    "questionary>=2.0.1",
    "anyio>=4.2.0",
    "jsonschema>=4.20.0",
    "python-dotenv>=1.0.0",
    # LLM Providers
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "tiktoken>=0.5.0",
    # Web Search
    "ddgs>=9.0.0",
    "beautifulsoup4>=4.12.0",
    # Document Processing
    "python-docx>=1.1.0",
    "pypdf>=4.0.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.23",
    "markitdown>=0.1.0",
    "Pillow>=10.0.0",
    # RAG Pipeline
    "haystack-ai>=2.0.0",
    "sentence-transformers>=2.2.2",
    # Data Processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "mypy>=1.7.0",
    "ruff>=0.1.9",
]
# Minimal install for CLI-only usage (no RAG/document processing)
minimal = [
    "click>=8.1.7",
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.43",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "httpx>=0.26.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
kautilya = "kautilya.cli:main"

[project.urls]
Homepage = "https://github.com/yourorg/agent-framework"
Documentation = "https://github.com/yourorg/agent-framework/docs"
Repository = "https://github.com/yourorg/agent-framework"

[tool.setuptools]
packages = ["kautilya", "kautilya.commands", "kautilya.templates", "kautilya.agent"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py311"
